% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlookup.R
\name{vlookup}
\alias{vlookup}
\title{use a key in one dataframe to look up a value in another}
\usage{
vlookup(this, df, key, value)
}
\arguments{
\item{this}{key in the current data frame}

\item{df}{dataframe with values of interest}

\item{key}{key that corresponds to 'this' in the dataframe 'df'}

\item{value}{value in the dataframe 'df' that is linked to 'key'}
}
\value{
dataframe
}
\description{
use a key in one dataframe to look up a value in another
}
